Restructuring of Trigger Checker for models
